Pull request 296: MIPS fixes
Merge in ADGUARD-CORE-LIBS/vpn-libs from mipswip to master

Squashed commit of the following:

commit 179abdaf8466605d6ec788f0902b1ba3d0e044ea
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Wed Mar 13 23:54:37 2024 +0200

    Fix wrong aliasing of port in SOCKS5 listener

commit 918ac35d1050ebca81da0a760a9ee42d37910c8e
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Wed Mar 13 22:27:09 2024 +0200

    Revert "wip"

    This reverts commit e20df9ece84a40f993af26be1e8a524c589b20f0.

commit 47330b7bae192c59b8f2db6b65db0c0cd2bf0c13
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Wed Mar 13 22:02:22 2024 +0200

    wip

commit d57ca2c341c839acafad44794a257db12d81fd0d
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Wed Mar 13 22:00:37 2024 +0200

    wip

commit f426154392b6d3c5595a94f43b42136faff223d3
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Wed Mar 13 21:55:47 2024 +0200

    wip

commit e1e765942aa90c26486a2fc2aec2bf9c35ae7cfc
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Wed Mar 13 21:27:19 2024 +0200

    wip

commit e20df9ece84a40f993af26be1e8a524c589b20f0
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Wed Mar 13 21:02:07 2024 +0200

    wip

+++ net/src/socks5_listener.cpp
@@ -483,7 +483,7 @@ static void complete_tcp_connection(Socks5Listener *listener, Connection *conn, 
     }
 
     sockaddr_storage local_addr = local_sockaddr_from_fd(tcp_socket_get_fd(conn->socket.get()));
-    int port = htons(sockaddr_get_port((struct sockaddr *) &local_addr));
+    uint16_t port = htons(sockaddr_get_port((struct sockaddr *) &local_addr));
     memcpy(reply->bnd_addr + offset, &port, 2);
 
     log_conn(listener, conn->id, conn->proto, dbg, "Sending reply");
@@ -633,7 +633,7 @@ int socks5_listener_send_data(Socks5Listener *listener, uint64_t id, const uint8
             break;
         }
 
-        int port = htons(sockaddr_get_port((struct sockaddr *) &dst->ip));
+        uint16_t port = htons(sockaddr_get_port((struct sockaddr *) &dst->ip));
         memcpy(reply->dst_addr + offset, &port, 2);
         offset += 2;
 
@@ -979,7 +979,7 @@ static bool complete_udp_association(Socks5Listener *listener, Connection *conn)
     reply->rsv = 0;
     reply->atyp = atyp;
 
-    int port = htons(sockaddr_get_port((struct sockaddr *) &bound_addr));
+    uint16_t port = htons(sockaddr_get_port((struct sockaddr *) &bound_addr));
     switch (bound_addr.ss_family) {
     case AF_UNSPEC:
         if (atyp == S5AT_IPV4) {
@@ -1458,7 +1458,7 @@ static void sock_handler(void *arg, TcpSocketEvent what, void *data) {
             reply->atyp = req->atyp;
             memcpy(reply->bnd_addr, req->dst_addr, addr_len);
             sockaddr_storage local_addr = local_sockaddr_from_fd(tcp_socket_get_fd(conn->socket.get()));
-            int port = htons(sockaddr_get_port((struct sockaddr *) &local_addr));
+            uint16_t port = htons(sockaddr_get_port((struct sockaddr *) &local_addr));
             memcpy(reply->bnd_addr + addr_len, &port, 2);
 
             error = tcp_socket_write(conn->socket.get(), (uint8_t *) reply_data, reply_size);
